+++ vite.config.js
@@ -9,8 +9,8 @@ export default defineConfig({
                 { src: path.resolve(__dirname, 'node_modules/jquery/dist/jquery.min.js'), dest: 'js' },
                 { src: path.resolve(__dirname, 'node_modules/jquery-ui/dist/jquery-ui.min.js'), dest: 'js' },
                 { src: path.resolve(__dirname, 'node_modules/bootstrap/dist/css/bootstrap.min.css*'), dest: 'css' },
-                { src: path.resolve(__dirname, 'node_modules/font-awesome/css/font-awesome.min.css'), dest: 'css' },
-                { src: path.resolve(__dirname, 'node_modules/font-awesome/fonts'), dest: '' },
+                { src: path.resolve(__dirname, 'node_modules/bootstrap-icons/font/bootstrap-icons.min.css'), dest: 'css' },
+                { src: path.resolve(__dirname, 'node_modules/bootstrap-icons/font/fonts'), dest: 'css' },
                 { src: path.resolve(__dirname, 'node_build/img/*'), dest: 'img' },
             ],
         })