+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Minimum PHP version is now 8.3, tested up to PHP 8.5
 - `MediaRepository` and `PreloadMediaRepository` uses `QueryBuilderFactoryInterface` instead of Connection
 - Removed Bootstrap 3 CSS classes from templates
+- Replaced Font Awesome with Bootstrap Icons
 
 ## [4.1.0] - 2025-11-10